% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_glmnet}
\alias{fit_glmnet}
\title{Fit a regularised logistic regression with 'glmnet'}
\usage{
fit_glmnet(x, y, folds, parallel = TRUE)
}
\description{
Fit a regularised logistic regression with 'glmnet'
}
