% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{hyper_search}
\alias{hyper_search}
\title{Perform a grid-search for the best hyper-parameters}
\usage{
hyper_search(candidates, x, y, folds, parallel)
}
\description{
Perform a grid-search for the best hyper-parameters
}
